Bump Arrow to 2.2.1.1 and simplify foldIndexed logic in Day 15

+++ src/main/kotlin/ru/timakden/aoc/year2023/Day15.kt
@@ -38,7 +38,7 @@ object Day15 {
         }
 
         return boxes.entries.sumOf { (boxIdx, lenses) ->
-            lenses.foldIndexed(0.toInt()) { lensIdx, acc, (label, focalLength) ->
+            lenses.foldIndexed(0) { lensIdx, acc, (label, focalLength) ->
                 acc + ((boxIdx + 1) * (lensIdx + 1) * focalLength)
             }
         }
